chore: move DataIntegrity
